Added my creation of 7-seg display app
The app helps in understanding the binary values generated to use them in different embedded projects. In the app, the user can click on segments and check the binary values needed to light up specific LEDs on 7-seg display. A simple button helps in checking values for both common anode and common cathode type 7-segment displays.